meson: define targets using data in lib/
There's only support for one target at the moment, but this
approach is likely to generalize well and is compatible with
libopencm3's existing tree structure.

The basic idea is to use dictionaries instead of subdir(). subdir()
doesn't provide fine enough control. The Make-based build system allows
each target to specify exactly which object files it includes. In
general, a target does not use all the object files from a subdir.
Therefore, Meson also has to use this level of precision. This is not
to preclude future optimizations using subdir(), but to ensure we have
an accurate build.