scripts/irq2nvic_h.py: accept path to input file
The previous version of this script constructed the path of the input
file based on assuming the value of @SOURCE_ROOT@ to be the root of
the libopencm3 project. When the project is used as a subproject, this
assumption is false, and the script fails because miscalculates the
path of the input file.

To fix this, let Meson calculate the path of the input file and let
the script accept that path.