meson: fix source generation
Meson treats files appearing in the sources list differently depending
on their extension. Files that end in .c are assumed to be valid compilation
units with a corresponding object file. Files that do not end in .c, which
includes .h files, are not treated as precursors to object files, but still
represent dependencies of a target, possibly generated. Therefore, Meson does
not support generating .c files which are not intended to be valid compilation
units themselves.

Of course, this is precisely the case with the vector_nvic.c for each target.
It is not a valid compilaton unit, but is included by another file which is.
Therefore, I changed the extension to suit Meson. The Make build system should
be able to adapt to Meson's enforced convention.

Note that other possible extensions are also valid, which may better convey
the intent to readers. vector_nvic.h is not a proper header file, but neither
is it a "proper" C file according to Meson. Perhaps it needs a different
extension to reflect this atypical status.